Namespace employer details by client
