--- app/forms/steps/income/employer_details_form.rb
@@ -1,21 +1,0 @@
-module Steps
-  module Income
-    class EmployerDetailsForm < Steps::BaseFormObject
-      attribute :employer_name
-      attribute :address
-
-      attribute :address_line_one
-      attribute :address_line_two
-      attribute :city
-      attribute :country
-      attribute :postcode
-
-      validates :employer_name, presence: true
-      validates_with AddressValidator
-
-      def persist!
-        record.update(attributes)
-      end
-    end
-  end
-end
